Feat : SpringBoot 회원가입 RESTful API 구현

+++ src/main/java/com/example/petcommunity/repository/member/MemberRepository.java
@@ -3,5 +3,5 @@ package com.example.petcommunity.repository.member;
 import com.example.petcommunity.entity.member.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRepository extends JpaRepository<MemberEntity, Long> {
+public interface MemberRepository extends JpaRepository<MemberEntity, Integer> {
 }
